+++ korus-ui/components/MaskedInput/theme.ts
@@ -1,5 +1,6 @@
 /* eslint-disable key-spacing */
 export const defaultMaskedInputTheme = {
+  closeIcon:                'masked-input-clear-icon',
   input:                    'masked-input-element',
   inputWrapper:             'masked-input-wrapper',
   inputWrapperDisabled:     'disabled',